requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
pdfkit==1.0.0
itsdangerous==2.1.2
email-validator==2.0.0
authlib==1.2.1